Rename: #12 레시피 페이지 폴더 생성 및 style.ts파일 생성
