+++ client/src/App.tsx
@@ -8,7 +8,7 @@ import Landing from './Pages/Landing';
 import LoginPage from './Pages/LoginPage/LoginPage';
 import SignUpPage from './Pages/SignUpPage/SignUpPage';
 import MyPage from './Pages/MyPage/MyPage';
-import RecipeList from './Pages/RecipeList';
+import RecipeList from './Pages/RecipePage/RecipeList';
 import RecipeWrite from './Pages/RecipeWrite';
 
 function App() {
